Adding `assign(count, item)` signature of assign. More to come to finish #55 Fixing #57. Breaking change but I doubt anybody is using these ctor signatures.
